Do not load template data from disk by default.
This change replaces the previously-used Makefile with go:generate
directives (for which go:generate was designed). It also alters the
default such that building an Appdash app will always produce a
binary that includes the template data.

This is easier for normal users, as majority of the time they are
not interested in modifying the template data. Below is a translation
of the make commands to their new equivalents:

  $ go generate

Will produce two files containing the template data. By default, the
data will be built into the binary for easy deployment etc by users.

  $ go install -tags=debug ./...

Building the code with the debug build tag will enable usage of the
second Go file generated, which loads the template data from disk
(useful for development of the template files).

Fixes #59